+++ src/Sulu/Bundle/ContentBundle/Resources/public/js/main.js
@@ -20,7 +20,8 @@ require.config({
         "type/block": '../../sulucontent/js/validation/types/block',
         "type/toggler": '../../sulucontent/js/validation/types/toggler',
         "extensions/sulu-buttons-contentbundle": '../../sulucontent/js/extensions/sulu-buttons',
-        "extensions/seo-tab": '../../sulucontent/js/extensions/seo-tab'
+        "extensions/seo-tab": '../../sulucontent/js/extensions/seo-tab',
+        "extensions/excerpt-tab": '../../sulucontent/js/extensions/excerpt-tab'
     }
 });
 
@@ -28,9 +29,10 @@ define([
     'config',
     'extensions/sulu-buttons-contentbundle',
     'extensions/seo-tab',
+    'extensions/excerpt-tab',
     'sulucontent/ckeditor/internal-link',
     'css!sulucontentcss/main'
-], function(Config, ContentButtons, SeoTab, InternalLinkPlugin) {
+], function(Config, ContentButtons, SeoTab, ExcerptTab, InternalLinkPlugin) {
     return {
 
         name: "Sulu Content Bundle",
@@ -40,6 +42,7 @@ define([
             'use strict';
 
             SeoTab.initialize(app);
+            ExcerptTab.initialize(app);
 
             var sandbox = app.sandbox;
             sandbox.sulu.buttons.push(ContentButtons.getButtons());